feat(primitives): expose HTTP retry utilities
