+++ crates/primitives/src/lib.rs
@@ -1,6 +1,8 @@
 //! Primitives for the taikoscope.
 /// Hardware cost estimates
 pub mod hardware;
+/// HTTP retry helpers
+pub mod http_retry;
 /// Simple rate limiter
 pub mod rate_limiter;
 /// Retry layer
